requirements.txt
MetaTrader5>=5.0.45
requests>=2.31
python-dotenv>=1.0
pandas>=2.0
numpy>=1.24
mplfinance>=0.12
matplotlib>=3.7
Pillow>=10.0
google-genai>=1.0
scipy>=1.11
